Add files via upload
Sectioned off some blocks of code, while also fixing spelling errors (adding commas and whatnot).
Changed $codename to $ubuntu, even though it could be "Debian", as Ubuntu is a subsection of Debian... ah, but that's a topic for another day. I've kept it as $ubuntu as to make it clear-er to those whom read it that may not have much experience, or knowledge, about Linux environments.